Home/LogIn

+++ SUS/Apps/Suls/Data/Problem.cs
@@ -19,7 +19,7 @@ namespace Suls.Data
         [MaxLength(20)]
         public string Name { get; set; }
 
-        public ushort Points { get; set; }
+        public int Points { get; set; }
 
         public virtual ICollection<Submission> Submissions { get; set; }
     }
